Move lockfile package types to package module
